src/scraper/scraper.py
# Scraper logic to fetch (requests) and parse the page (BeautifulSoup) then extract the data from structured HTML
import requests
from bs4 import BeautifulSoup

def fetch_page(url):
    response = requests.get(url)
    return response.text if response.status_code == 200 else None

def parse_page(html):
    soup = BeautifulSoup(html, 'html.parser')
    auctions = soup.find_all('tr', {'data-target': 'sale-details'})  # Adjust the selector as per actual page
    data = []
    for auction in auctions:
        case_number = auction.find('td', {'data-label': 'Case #'}).text.strip()
        sale_date = auction.find('td', {'data-label': 'Sale Date'}).text.strip()
        property_address = auction.find('td', {'data-label': 'Property Address'}).text.strip()
        status = auction.find('td', {'data-label': 'Status'}).text.strip()
        data.append({
            'case_number': case_number,
            'sale_date': sale_date,
            'property_address': property_address,
            'status': status
        })
    return data
